pyproject.toml
[project]
name = "hermes-a2a"
version = "0.1.0"
description = "A2A protocol bridge to the local Hermes agent"
requires-python = ">=3.11"

dependencies = [
    "a2a-sdk>=0.3.25",
    "httpx>=0.28.1",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.1.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.9.0",
    "starlette>=0.45.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "respx>=0.22.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hermes_a2a"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
